DEV : Add some properties and relations in the entities

+++ BCM/src/main/java/org/example/bcm/core/model/entity/Challenge.java
@@ -30,5 +30,5 @@ public class Challenge {
     @JoinColumn(name = "table_id")
     private Table table;
 
-    private Long numberOfParties;
+    private Integer numberOfParties;
 }
